:bug: FIX: One command per line and cammelCase enforced (because of requirement changes)
