+++ config/index.go
@@ -5,9 +5,11 @@ const AmntBitsParam = 8
 
 var Testing bool = false
 
-const VariableNameRegex = "^[a-zA-Z][a-zA-Z0-9_]*$"
+const VariableNameRegex = "^[a-z][a-zA-Z0-9]*$"
 
 const AssemblyFileExtension = ".asm"
 
-const SuccessStatus = 0
-const FailStatus = 1
+const (
+	SuccessStatus = 0
+	FailStatus    = 1
+)
