+++ core/label/label_test.go
@@ -16,22 +16,22 @@ func TestAssembleJumpLabel(t *testing.T) {
 		// Success with label and rest
 		{"label:mov 12", helper.StringPointer("label"), "mov 12", false},
 		{"label :mov 12", helper.StringPointer("label"), "mov 12", false},
-		{"label: mov 12", helper.StringPointer("label"), "mov 12", false},
-		{"label : mov 12", helper.StringPointer("label"), "mov 12", false},
+		{"label: mov 12", helper.StringPointer("label"), " mov 12", false},
+		{"label : mov 12", helper.StringPointer("label"), " mov 12", false},
 		// Success without label
 		{"mov 12", nil, "mov 12", false},
 		// Success without rest
 		{"label:", helper.StringPointer("label"), "", false},
 		{"label :", helper.StringPointer("label"), "", false},
-		{"label: ", helper.StringPointer("label"), "", false},
-		{"label : ", helper.StringPointer("label"), "", false},
+		{"label: ", helper.StringPointer("label"), " ", false},
+		{"label : ", helper.StringPointer("label"), " ", false},
 		// Fail: Invalid name
-		{"1label : mov 12", nil, "mov 12", true},
+		{"1label : mov 12", nil, " mov 12", true},
 		{".label :mov 12", nil, "mov 12", true},
-		{": mov 12", nil, "mov 12", true},
+		{": mov 12", nil, " mov 12", true},
 		// Multiple colons
-		{": mov: 12", nil, "mov: 12", true},
-		{"label: mov: 12", helper.StringPointer("label"), "mov: 12", false},
+		{": mov: 12", nil, " mov: 12", true},
+		{"label: mov: 12", helper.StringPointer("label"), " mov: 12", false},
 	}
 
 	for i, test := range tests {
@@ -47,7 +47,7 @@ func TestAssembleJumpLabel(t *testing.T) {
 		}
 
 		if gotRestOfLine != test.expectedRestOfLine {
-			t.Errorf("[%d] Expected: %s, Got: %s // param: '%s'", i, test.expectedRestOfLine, gotRestOfLine, test.param)
+			t.Errorf("[%d] Expected: '%s', Got: '%s' // param: '%s'", i, test.expectedRestOfLine, gotRestOfLine, test.param)
 		}
 	}
 }
