+++ utils/validations.go
@@ -1,9 +1,10 @@
 package utils
 
 import (
-	"github.com/open-machine/assembler/config"
 	"math"
 	"regexp"
+
+	"github.com/open-machine/assembler/config"
 )
 
 func IsOverflow(num uint, availableBits int) bool {
@@ -15,3 +16,8 @@ func IsValidVarName(str string) bool {
 	matched, err := regexp.MatchString(config.VariableNameRegex, str)
 	return matched && err == nil
 }
+
+// TODO: check valid name
+// TODO: use to validate name
+// func isReservedWord(str string) bool {
+// }
